mysite_com/requirements.txt
whitenoise 
dj-database-url 
django-hosts
django-environ
pytz 
asgiref 
sqlparse 
Django 
django-heroku 
gunicorn 
numpy 
six 
python-dateutil 
pandas 
python-dotenv 
tqdm 
regex 
joblib 
click 
nltk 
opencv-python 
charset-normalizer 
idna 
certifi 
urllib3 
requests 
Pillow 
pytesseract 
greenlet 
SQLAlchemy